# Fix missing cache lookups