… .editorconfig
… tests/Directory.Build.targets
… tests/ImageSharp.Web.Benchmarks/MemoryConfig.cs
… tests/ImageSharp.Web.Tests/Caching/PhysicialFileSystemCacheTests.cs
… tests/ImageSharp.Web.Tests/Commands/CommandParserTests.cs
… tests/ImageSharp.Web.Tests/ImageSharp.Web.Tests.csproj