### Prerequisites

- [x] I have written a descriptive pull-request title
- [x] I have verified that there are no overlapping [pull-requests](https://github.com/SixLabors/ImageSharp.Web/pulls) open
- [x] I have verified that I am following matches the existing coding patterns and practice as demonstrated in the repository. These follow strict Stylecop rules :cop:.
- [x] I have provided test coverage for my change (where applicable)

### Description

Fixes https://github.com/SixLabors/ImageSharp.Web/issues/140

Simple option. 

Merge them into one cache, and treat them as tuples.

Other option is to cache it locally inside `IImageCacheResolver` so that subsequent calls retrieve the same instance.

But I think this involves less changes.