@JimBobSquarePants 2021-02-26
Code Coverage drop is caused by a binding redirect to a buggy version of Coverlet. Updating to the latest submodule and building the solution should fix this