@JimBobSquarePants 2021-02-26
@deanmarcussen Lemme pull down your branch. I think there mhave been a breaking change in the submodule paths causing the stylecop rule changes