@JimBobSquarePants 2021-02-26
No worries mate. Looks like the xunit upgrade has busted `CommandParserCanConvert<T>(T expected, string param, CultureInfo culture)` unit tests. I'm getting ambiguous match exceptions. 

xunit test thyself!