@sebastienros 2021-02-26
When I sent my previous comment, I was afraid that stylecop would refuse it because it has a slash in it.