@deanmarcussen 2021-02-26
Might have won a little bit from it as well.
`http://localhost:2646/imagesharp-logo.png?width=300`

before:
```
Statistics        Avg      Stdev        Max
  Reqs/sec     40457.76    9068.64   54989.01
  Latency        2.47ms     1.90ms   129.56ms
  HTTP codes:
    1xx - 0, 2xx - 805425, 3xx - 0, 4xx - 0, 5xx - 3552
    others - 0
  Throughput:   191.02MB/s
```

after:

```
Statistics        Avg      Stdev        Max
  Reqs/sec     42820.66    6498.54   55995.16
  Latency        2.33ms   689.17us   221.96ms
  HTTP codes:
    1xx - 0, 2xx - 856105, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:   203.31MB/s
```