@sebastienros 2021-02-26
How can we measure the max speed of a disk. Wondering if the limit here is how fast it can read the file. Totally unrelated to the PR.