@sebastienros 2021-02-26
Apparently on windows you can do:

```
winsat disk -drive c

> Disk  Random 16.0 Read                       789.24 MB/s          8.6
> Disk  Sequential 64.0 Read                   1433.71 MB/s          8.7
> Disk  Sequential 64.0 Write                  646.15 MB/s          8.2
```