@JimBobSquarePants 2021-02-27
I'm gonna merge this and also produce a stable release for NuGet to make it easier to consume. Thanks for the assist!